+++ R/layer_helpers.R
@@ -18,7 +18,7 @@
 #'   stat = "my_custom_stat",
 #'   position = "identity",
 #'   ...,
-#'   inset = NA,
+#'   inset = ggplot2::waiver(),
 #'   map_base = "normal",
 #'   map_inset = "auto",
 #'   na.rm = TRUE,
@@ -26,8 +26,10 @@
 #' ) {
 #'   params <- rlang::list2(na.rm = na.rm, ...)
 #'   build_sf_inset_layers(
-#'     data = data, mapping = mapping,
-#'     stat = stat, position = position,
+#'     data = data,
+#'     mapping = mapping,
+#'     stat = stat,
+#'     position = position,
 #'     show.legend = show.legend,
 #'     inherit.aes = inherit.aes,
 #'     params = params,
@@ -37,23 +39,42 @@
 #'   )
 #' }
 build_sf_inset_layers <- function(
-  data, mapping, stat, position, show.legend, inherit.aes, params, inset,
-  map_base = "normal", map_inset = "auto"
+  data,
+  mapping,
+  stat,
+  position,
+  show.legend,
+  inherit.aes,
+  params,
+  inset,
+  map_base = "normal",
+  map_inset = "auto"
 ) {
+  if (!is.null(inset) && length(inset) == 1L && is.na(inset)) {
+    lifecycle::deprecate_warn(
+      when = "0.5.0",
+      what = "build_sf_inset_layers(inset = 'must not be NA')",
+      details = "use `inset = waiver()` instead to default to the coord's inset"
+    )
+    inset <- waiver()
+  }
   has_inset_cfg <- !is.null(inset)
 
   map_base <- rlang::arg_match0(map_base, c("normal", "clip", "none"))
   map_inset <- rlang::arg_match0(map_inset, c("auto", "normal", "none"))
 
   if (map_inset == "auto") {
-    draw_inset <- map_inset == "normal" || (map_inset == "auto" && has_inset_cfg)
+    draw_inset <- map_inset == "normal" ||
+      (map_inset == "auto" && has_inset_cfg)
     map_inset <- if (draw_inset) "normal" else "none"
   }
   if (map_base == "none" && map_inset == "none") {
-    cli::cli_abort("{.arg map_base} and {.arg map_inset} can't both be disabled")
+    cli::cli_abort(
+      "{.arg map_base} and {.arg map_inset} can't both be disabled"
+    )
   }
 
-  make_layer <- function(inset_mode) {
+  make_layer <- function(inset_mode, layer_type) {
     ggplot2::layer_sf(
       data = data,
       mapping = mapping,
@@ -62,24 +83,38 @@ build_sf_inset_layers <- function(
       position = position,
       show.legend = show.legend,
       inherit.aes = inherit.aes,
-      params = modifyList(params, list(inset = inset, inset_mode = inset_mode))
+      params = modifyList(
+        params,
+        list(inset = inset, inset_mode = inset_mode, layer_type = layer_type)
+      )
     )
   }
 
-  base_layer <- switch(map_base,
-                       normal = make_layer("none"),
-                       clip = make_layer("cutout"),
-                       none = NULL)
-  inset_layer <- switch(map_inset,
-                        normal = make_layer("normal"),
-                        none = NULL)
+  # There are 3 possibilities for each of the two layers:
+  #   1) we know it needs to be drawn (map_base/map_inset not "none"), or
+  #   2) we know it should not be drawn (map_base/map_inset is "none"), or
+  #   3) we don't know either way because the inset configuration is deferred
+  #      to the coord and will be computed later.
+  # For 2) we can just set the layer to NULL to avoid adding an extra layer to
+  # the plot. For 3) we have to add the layer but allow for it to be a no-op.
+  base_layer <- switch(
+    map_base,
+    normal = make_layer("none", "base"),
+    clip = make_layer("cutout", "base"),
+    none = NULL
+  )
+  inset_layer <- switch(
+    map_inset,
+    normal = make_layer("normal", "inset"),
+    none = NULL
+  )
   c(base_layer, inset_layer, ggplot2::coord_sf(default = TRUE))
 }
 
 #' Get the inset configuration from the params or coord
 #'
 #' This is a helper for implementing inset-aware ggplot layers. If the `inset` is
-#' missing (`NA`) then the default inset configuration is retrieved from the coord.
+#' not specified then it is retrieved from the coord if present.
 #'
 #' @param inset Inset passed in as a param to the layer
 #' @param coord Coord object for the plot
@@ -89,7 +124,7 @@ build_sf_inset_layers <- function(
 #' @examples
 #' # defining a new geom deriving from geom_sf()
 #' GeomCustom <- ggplot2::ggproto("GeomCustom", ggplot2::GeomSf,
-#'   draw_panel = function(self, data, panel_params, coord, inset = NA) {
+#'   draw_panel = function(self, data, panel_params, coord, inset = ggplot2::waiver()) {
 #'     inset <- get_inset_config(inset, coord)
 #'
 #'     # do something with the inset ...
@@ -102,7 +137,7 @@ build_sf_inset_layers <- function(
 get_inset_config <- function(inset, coord) {
   if (is.null(inset)) {
     NULL
-  } else if (is.na(inset)) {
+  } else if (is_waiver(inset)) {
     if (inherits(coord, "CoordSfInset")) {
       make_inset_config(coord$inset)
     } else {